requirements.txt
git+https://github.com/openai/whisper.git
torch
torchaudio
langchain
langchain-core
langchain-community
langchain-groq
speechrecognition
sounddevice
scipy
python-dotenv
huggingface_hub
pyttsx3
qrcode
Pillow
pyaudio